fit 2d gaussian for laser intensity
the fit is done for laser intensity. Using WaveMax() was sometimes
leading to bad results. As we are recording the laser focus
there should only be one visible gauss profile in the image. Using a
simple 2d gaussian should du the job. FWHM could also be used to find
the maximum of the laser.