add procedure to combine two coordinate lists
if two coordinate waves from diffrent images have to get merged to a
single wave we can not use the builtin concatenate for it since it is
only aware of 1D waves. this procedures can be used to concatenate two
coordinate waves. Maybe it is not suitable to have this accessible
from the menu as it is right now.