add path to experiments
The sma package stores a list of source files for SMAread() to read in and send to PLEMd2open().

This list was stored as full absolute path. any change in directory structure did affect the finding of files.

This update now stores/converts all absolute paths to relative paths. A overall package path is introduced that is used as root directory for all paths.